+++ packages/dag/package.json
@@ -20,7 +20,7 @@
     "react-dom": "^16.0.0"
   },
   "devDependencies": {
-    "@latticejs/utils": "file:../utils",
+    "@latticejs/storybook-readme": "^1.0.1-alpha.7",
     "@material-ui/core": "^1.0.0",
     "@storybook/addon-actions": "^3.4.10",
     "@storybook/addon-storyshots": "^3.4.5",
