+++ packages/infinite-list/package.json
@@ -25,7 +25,7 @@
     "react-dom": "^16.0.0"
   },
   "devDependencies": {
-    "@latticejs/utils": "file:../utils",
+    "@latticejs/storybook-readme": "^1.0.1-alpha.7",
     "@material-ui/core": "^1.0.0",
     "@material-ui/icons": "^2.0.0",
     "@storybook/addon-storyshots": "^3.4.5",
