+++ packages/infinite-list/stories/list.js
@@ -10,7 +10,7 @@ import { List, ListItem } from '../src';
 import muiTheme from '../.storybook/decorator-material-ui';
 import Readme from '../README.md';
 import { JssDecorator } from './utils';
-import { withReadme } from '@latticejs/utils';
+import { withReadme } from '@latticejs/storybook-readme';
 
 // Decorators
