+++ packages/tree/stories/tree.js
@@ -4,7 +4,7 @@ import { action } from '@storybook/addon-actions';
 // Ours
 import Tree from '../src/tree';
 import muiTheme from '../.storybook/decorator-material-ui';
-import { withReadme } from '@latticejs/utils';
+import { withReadme } from '@latticejs/storybook-readme';
 import Readme from '../README.md';
 
 const Flexed = story => (
