--- packages/utils/src/main.js
@@ -1,1 +1,0 @@
-export * from './stories';
